Move webapp to liaison
